Fixed code after feedback
